+++ include/random.h
@@ -2,9 +2,9 @@
 #define RANDOM_H
 
 
-void fill_random(float a, float b, float* array, int size);
-float rand01();
-void fill_population(float* positions, int N, int n, float* bounds[]);
+void fill_random(double a, double b, double* array, int size);
+double rand01();
+void fill_population(double* positions, int N, int n, double* bounds[]);
 int randint(int a, int b);
 void randperm(int* values, int len);
