+++ src/random.cpp
@@ -1,29 +1,29 @@
 #include <stdlib.h>
 #include <time.h>
 
-void fill_random(float a, float b, float* array, int size)
+void fill_random(double a, double b, double* array, int size)
 {
-    float l = b-a;
+    double l = b-a;
     int i;
 
     for (i = 0; i < size; ++i) {
-        array[i] = a + l * (float) rand() / (float)RAND_MAX ;
+        array[i] = a + l * (double) rand() / (double)RAND_MAX ;
     }
     
 }
 
 
-float rand01()
+double rand01()
 {
-    return (float) rand() / (float)RAND_MAX;
+    return (double) rand() / (double)RAND_MAX;
 }
 
 
 
-void fill_population(float* positions, int N, int n, float* bounds[])
+void fill_population(double* positions, int N, int n, double* bounds[])
 {
     int i, j;
-    float l;
+    double l;
     for (j = 0; j < n; ++j) {
         l = bounds[j][1] - bounds[j][0];
         for (i = 0; i < N; ++i) {
